+++ changelog.md
@@ -23,6 +23,7 @@
       - Remove buckler as an accessory for armors
   - Progress dialog: fix task bar progress not being cleared on close
   - Status: Attempt to fix weather overlay again
+  - Display: Set line-height to better control unicode
 
 ## v0.6.8 2018-11-12