+++ changelog.md
@@ -24,6 +24,7 @@
       - Do not generate set_prevent_* functions of empty string for objects
       - Cancel on error when generating code
       - Remove buckler as an accessory for armors
+      - Fixed set_long string wrapping
   - Progress dialog: fix task bar progress not being cleared on close
   - Status: Attempt to fix weather overlay again
   - Display: Set line-height to better control unicode