+++ changelog.md
@@ -27,6 +27,7 @@
       - Fixed set_long string wrapping
       - Fixed room set_long/short when empty
       - Fixed monster/object file name if empty
+      - Fixed several minor issues with code generation
   - Progress dialog: fix task bar progress not being cleared on close
   - Status: Attempt to fix weather overlay again
   - Display: Set line-height to better control unicode